requirements.txt
flask
line-bot-sdk
openai>=1.0.0
python-dotenv